Moved db into its own package
