--- packages/api/.prettierignore
@@ -1,1 +1,0 @@
-src/db/migrations
