+++ packages/db/drizzle.config.ts
@@ -1,9 +1,9 @@
 import { defineConfig } from 'drizzle-kit';
-import { getDBConnectionString } from './src/db/connection';
+import { getDBConnectionString } from './src/connection';
 
 export default defineConfig({
-  schema: './src/db/schema/**/*.ts',
-  out: './src/db/migrations',
+  schema: './src/schema/**/*.ts',
+  out: './src/migrations',
   driver: 'pg',
   dbCredentials: {
     connectionString: getDBConnectionString(),
