+++ packages/db/example.env
@@ -1,7 +1,7 @@
 # Example '.env' file, copy the contents of this file to your own '.env'
 
-API_PORT=8080
 ENVIRONMENT=dev
+POSTGRES_PASSWORD=dev
 DB_PASSWORD=dev
 DB_USER=postgres
 DB_HOST=localhost
