+++ packages/db/src/schema/userSessions.ts
